requirements.txt
Package                      Version
---------------------------- ---------
annotated-types              0.7.0
anyio                        4.12.0
blinker                      1.9.0
cachetools                   6.2.4
certifi                      2025.8.3
charset-normalizer           3.4.2
click                        8.2.1
colorama                     0.4.6
distro                       1.9.0
filelock                     3.18.0
Flask                        3.1.1
fsspec                       2025.7.0
google-ai-generativelanguage 0.6.15
google-api-core              2.28.1
google-api-python-client     2.187.0
google-auth                  2.45.0
google-auth-httplib2         0.3.0
google-genai                 1.56.0
googleapis-common-protos     1.72.0
grpcio                       1.76.0
grpcio-status                1.71.2
h11                          0.16.0
httpcore                     1.0.9
httplib2                     0.31.0
httpx                        0.28.1
huggingface-hub              0.34.3
idna                         3.10
itsdangerous                 2.2.0
Jinja2                       3.1.6
joblib                       1.5.1
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
mdurl                        0.1.2
mpmath                       1.3.0
networkx                     3.5
numpy                        2.3.2
packaging                    25.0
pillow                       11.3.0
pip                          25.1.1
proto-plus                   1.27.0
protobuf                     5.29.5
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.12.5
pydantic_core                2.41.5
Pygments                     2.19.2
pyparsing                    3.3.1
PyYAML                       6.0.2
regex                        2025.7.34
requests                     2.32.4
rich                         14.1.0
rsa                          4.9.1
safetensors                  0.5.3
scikit-learn                 1.7.1
scipy                        1.16.1
sentence-transformers        5.0.0
setuptools                   80.9.0
sniffio                      1.3.1
sympy                        1.14.0
tenacity                     9.1.2
threadpoolctl                3.6.0
tokenizers                   0.21.4
torch                        2.7.1
tqdm                         4.67.1
transformers                 4.54.1
typing_extensions            4.14.1
typing-inspection            0.4.2
uritemplate                  4.2.0
urllib3                      2.5.0
websockets                   15.0.1
Werkzeug                     3.1.3
